Recibe claves el programa permutacion, pero no busca archivo o utiliza n

+++ Practica1_P2_G1242/src/algoritmos.c
@@ -469,7 +469,7 @@ void free_mpz_matrix(mpz_t** mat, int rows, int cols) {
 
 long double average(int how_many, long double* vals) {
 	int i = 0;
-	double sum = 0;
+	long double sum = 0;
 	
 	for (i = 0; i < how_many; i++) {
 		sum += vals[i];
